Safari: fix product grid layout (#1637)
## Describe your changes

Fix Safari CSS grid bug in product grid.

![Screenshot 2023-02-13 at 18.18.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/fc0ff227-9fb7-4ee0-8cea-87397205ddb4/Screenshot%202023-02-13%20at%2018.18.04.png)

## Justify why they are needed

CSS grid with `align-items: baseline` isn't working in Safari. This is a known bug: https://stackoverflow.com/questions/52502989/css-grid-align-items-baseline-not-working-in-safari.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code