SSR: update how to get hedvig-language header (#1645)
## Describe your changes

Update how to get hedvig-language header on server.

Skip relying on fallback.

## Justify why they are needed

We sometimes fail to determine language on the server. Then we fallback to english (SE) which is not good. Half the UI is in english and half in swedish.

I would like to refactor more and get rid of the fallback but I think this is a good first step.

The language should always be derived from the URL.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code