Add hover state to edit price calculator (#1650)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add hover state to edit button in price calculator.

![Screenshot 2023-02-14 at 15.41.29.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/ec6e8915-ffb5-40be-ba48-e0a4c7b116c7/Screenshot%202023-02-14%20at%2015.41.29.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code