Video download size optimizations (#1669)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Fix invalid style when video poster is missing
- Don't preload full videos by default

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Reduce initial download size of front page

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code