Add custom 500 error page (#1667)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Add custom error 500 page
- Refactor 404 error page and create reusable error page component

![Screenshot 2023-02-15 at 13.20.42.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/9c010f00-52a1-406b-9173-7f61de5be6f7/Screenshot%202023-02-15%20at%2013.20.42.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

They are basically the same page.

Caveat: it's not 100% clear what should happen when user clicks "try again".

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code