Predefine different page/block layouts (#1670)
## Describe your changes

Setup pre-defined layouts based on the 12 column grid.

We have 4 pre-defined widths:

- full (1)
- two-thirds (2/3)
- half (1/2)
- third (1/3)

We have 3 pre-defined alignments:

- left
- center
- right

## Justify why they are needed

This will make it easier to create layouts that are consistent with the design system. They will also behave the same across breakpoints.

The idea is to control these settings from the CMS on a block level.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code