Update details sheet layout (#1675)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Use grid layout for cart details sheet.

Clear the modal footer inside the dialog.

![Screenshot 2023-02-16 at 10.58.53.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/8a29e569-ffa3-4e44-ba14-916f7c9dded1/Screenshot%202023-02-16%20at%2010.58.53.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code