Optimize video posters through Storyblok optimization server (#1674)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Use Storyblok image optimization for video posters

This reduces resources dramatically, ~15MB total savings on /se, for example.  I'd love to use Vercel for this optimization, but they don't have a documented API outside of Image component

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code