Differentiate between product card link types (#1676)
## Describe your changes

Differentiate between product cards that lead to product page (should have extra get price link) and those that lead to landing page (content).

## Justify why they are needed

We don't get the content type from the CMS so we can't base it on that.

However, we could match against the API products. That requires another context though and I would rather test out Jotai for this.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code