Add very short variant display names to comparison table (#1683)
## Describe your changes

Add the very short variant display names to the comparison table (car).

Factor out the table into a separate component.

## Justify why they are needed

`OfferPresenter` is getting too big.

We can't fetch the display names from the API yet so I think we should hardcode them for now.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code