Quick add to cart: scroll to top of page (#1786)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update quick "add to cart" button on the footer of PDP page (mobile only)

Scroll all the way to the top of the page when clicking on the "add to cart" button

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Before it just scrolled to the main "add to cart" button, which looks a bit weird.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code