Add feature flag for english locale (#1791)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add feature flag for enabling/disabling the english locale.

Hide language selector if English locale is not available.



![Screenshot 2023-02-27 at 15.17.50.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/b3b6855a-69f2-41f1-b834-8e88b7b617e3/Screenshot%202023-02-27%20at%2015.17.50.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

For the launch.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code