Various menu tweaks (#1789)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Make top menu item clickable (links to store page)
- Remove "försäkringar" heading from product navigation section
- Render product navigation section outside menu item so it never closes



![Screenshot 2023-02-27 at 10.06.38.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/ff0cb0fb-6fa7-4c00-ab72-b9aab01d880f/Screenshot%202023-02-27%20at%2010.06.38.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

I couldn't figure out a way to avoid using "useBreakpoint". The more I work with this component, the more I think we should separate it into two components. One for mobile and one for desktop. It's getting too complicated.

Perhaps we can do that as a mob programming session?

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code