Disable cache version handling in Storyblok (#1801)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Disable manual cache version handling temporarily.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We are seeing some strange issues were some people are not able to view the site but others are.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code