Update styles for tier selector (#1811)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update styles for tier selector to match design.

[Screen Recording 2023-02-28 at 14.00.44.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/615ea2ea-10ed-4199-8c1f-d46da9d4d72f/Screen%20Recording%202023-02-28%20at%2014.00.44.mov)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Launch is soooooooooooooooooooooon!

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code