Make main menu item clickable/link (#1814)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Make the primary menu item clickable - link to the first page of the current section.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Requested from design.

Do we think this implicit solution is good enough? Or should we add a new field to the CMS for this?

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code