Update shopping cart icon to match pillow radius (#1819)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes



![Screenshot 2023-02-28 at 17.11.16.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/9c69f565-4898-4565-b590-ee16dd4a95fd/Screenshot%202023-02-28%20at%2017.11.16.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code