Update pillow placeholder image (#1824)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes



![Screenshot 2023-03-01 at 09.07.56.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/7c5890c9-e9d1-48e7-97aa-08e6620918b0/Screenshot%202023-03-01%20at%2009.07.56.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code