Add feature flag for country selector (#1835)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add feature flag to be able to toggle country selector in footer.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

It doesn't work linking between domains.

Peter says we also don't need it for now.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code