Add grid layout to perils block (#1840)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Integrate grid layout in perils block.

![Screenshot 2023-03-01 at 16.34.09.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/16d73baf-5bdb-4f1d-9471-132a2f5556a8/Screenshot%202023-03-01%20at%2016.34.09.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Fix an issue where perils block doesn't have the same max width as rest of blocks.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code