Campaign API route: always redirect to user defined URL (#1863)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Always redirect to the user defined path after attaching campaign code.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Before we used the fallback path if something went wrong but this is unexpected behaviour.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code