Disable remote cache for onboarding app (#1889)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Disable remove cache for onboarding app

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We don't really need it.

We need to do it in code since you can only disable it on the team level.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code