Allow app to recover from network errors while signing (#1892)
## Describe your changes

Differentiate between API/Server errors and network errors.

Ignore network errors when signing and keep polling for status.

## Justify why they are needed

We failed as soon as we got a network error but we should keep polling for status and only fail if API tells us the request failed.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code