Update to latest Next.js version (#1897)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update to latest version of Next.js

Replace `@next/font` with built-in alternative.

Refactor app-dir emotion setup according to new Next.js docs.

Fix bug where we redirect from 404 page.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

The debugger page styles are sort of broken but this Next.js release includes a bunch of fixes for the production site.

There seems to be serious issue with using emotion with app-dir/server components. Let's discuss this in the group!

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code