Add combobox component based on downshift (#1912)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add combobox component (autocomplete).


![Screenshot 2023-03-09 at 09.02.41.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/5680dfc7-1ea4-4f72-bc78-6beaa35505da/Screenshot%202023-03-09%20at%2009.02.41.png)

## Justify why they are needed

This will be needed for breed selector to launch PET.

Based on downshift lib. We don't want to implement all the accessibility stuff ourselves.

I investigated Radix but they don't offer a combobox so we would have to do a lot of wiring ourselves. I also tried out Reach UI but it wasn't working 100% they way we wanted. It has has also not been updated in 6 months which is a red flag.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code