GRW-2348 / Exclude pages from google indexing (#1898)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Added a robot.txt file with the pages to exclude. I have never done this before and am not sure how to test it so please feedback if something looks weird! 🤗

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): [GRW-2348]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-2348]: https://hedvig.atlassian.net/browse/GRW-2348?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ