Update date input component in purchase form (#1918)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Use a shared date input component for both purchase form and cancellation form.


![Screenshot 2023-03-10 at 10.37.06.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/0111046c-20e6-4564-8a1a-75470ac6fe7e/Screenshot%202023-03-10%20at%2010.37.06.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Replace the old date input with the new one.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code