Fix TextField warnings (#1942)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix 2 React warnings from TextField
- Do not pass onValueChange to DOM
- Do not pass defaultValue since we're using fully controlled component

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code