Update switching assistant to work on-demand (#1891)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Use new banksignering approval mutation to initiate switching flow
- Fix how checkout steps are calculated
- Fix redirect after successful payment connection
- Refactor switching assistant to include shop session ID in the URL

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Properly support BankSignering flow.

Switching is post-sign so we will have created a new "current" shop session. Therefore we need to keep track of the shop session you are "checking out" by keeping the ID in the URL.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code