Bug fix: running multiple confirm on price intent (#1969)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Run two mutations (`priceIntentUpdate` + `shopSessionCustomerUpdate`) at the same time in the price calculator.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Fix a bug where we try to confirm a price intent twice simultaneously. This happened because we are responding to two `onComplete` events from the individual mutations. This makes it so that we only get a single `onComplete` event.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code