Reset auth when resetting session (#1992)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Reset auth tokens when resetting session after wanting to change SSN.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Otherwise we just create a new *authenticated* shop session so it retains the same SSN 🙃.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code