Fix bug in parsing user errors (#1993)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix bug in parsing user error messages from backend.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We didn't account for the fact that "result" variable can be `null` / `undefined` so we got a runtime error.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code