Fix borders in cart inventory / cart page (#1997)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix some minor design issues with the cart inventory.


![Screenshot 2023-03-20 at 11.20.41.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/aaa2ecab-2db1-497d-853e-4069e265dc76/Screenshot%202023-03-20%20at%2011.20.41.png)


![Screenshot 2023-03-20 at 11.20.51.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/d1ba7695-b94e-4a88-8a91-9a54fe287863/Screenshot%202023-03-20%20at%2011.20.51.png)


![Screenshot 2023-03-20 at 11.21.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/49fe5975-5e0d-4a5a-acb0-885ed66c36e1/Screenshot%202023-03-20%20at%2011.21.04.png)


![Screenshot 2023-03-20 at 11.21.17.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/fcf6b811-41d1-4162-85cd-c5b63ecfdc93/Screenshot%202023-03-20%20at%2011.21.17.png)

Add hover state to delete buttons.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Feedback from Wille.

I thought we were already using the `CartInventory` component in multiple places but we are not. We should do that update soon.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code