Lock scroll position for purchase form (#2006)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Lock the scroll position of purchase form so it never scrolls with the rest of the content.

- Remove the top padding of the PDP video.


![Screenshot 2023-03-21 at 09.14.09.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/0769470c-a300-45b9-8b2c-2c97a24a2aa1/Screenshot%202023-03-21%20at%2009.14.09.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

UI improvements from Wille.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code