Bug: keep track of latest selected offer on PDP (#2010)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Refresh selected offer after editing price intent/offer (start date or re-confirm).

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We have a bug now where you don't see the updated start date after you edit. This is because we store the selected offer in local state and don't update it after editing. This PR fixes that.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code