Add page transition progress indicator (#2013)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Show the progress indicator on the top of the page during page transitions.

## Justify why they are needed

To make the page transitions more smooth.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code