Make Hedvig-logo link to home in Checkout (#2016)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Make Hedvig-logo link to home page in Checkout

- Align height of menu in Checkout with the rest of the site

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Request from Design.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code