Redirect to cart page after removing last entry (#2018)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Redirect the user to the cart page if they remove the last item in the cart on the Checkout page.

## Justify why they are needed

This is a bug fix.

It would be nice to avoid prop-drilling. I was considering some Jotai solution but not sure how it would work. I was also considering sending some event you can subscribe to but maybe this is still the easiest.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code