Update combobox design with shadow/zindex (#2042)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Add shadow to combobox when expanded
- Add z-index to combobox when expanded
- Increase font size of combobox
- Update padding in mixed breed picker
- Move up breed picker for dog (and cat?)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Improve UX and match design.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code