Update details sheets for PET insurance (#2054)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update details sheets for pet insurance.


![Screenshot 2023-03-27 at 15.31.18.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/c2da1339-5be5-4488-9169-07600b1c2be5/Screenshot%202023-03-27%20at%2015.31.18.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code