Use new API for BankSignering Switching (#2020)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Use new BankSignering API.
- Always send registration number without whitespace to PriceIntent API.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Switch to the API powered by Storefront.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code