Show warning icon + message in text field (#2061)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Show warning icon in text field alongside message.
- Increase space between text field and button in SSN input.


![Screenshot 2023-03-28 at 17.19.14.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/a3336ba7-e1a9-4d2c-87f1-b045a45906ab/Screenshot%202023-03-28%20at%2017.19.14.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

UI feedback from Wille.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code