Revert media query hook optimization (#2066)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Revert optimization to `useMediaQuery` hook.


![Screenshot 2023-03-29 at 14.15.51.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/ca1fb1ef-3e45-4a54-9377-a8cce89f1745/Screenshot%202023-03-29%20at%2014.15.51.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Was causing visual bugs in Perils component.

It also didn't resolve the original issue.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code