Implement global banner component (#2065)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Implement global banner component

- Support single use case: success banner after adding campaign code

- Show banner after adding forever code


![Screenshot 2023-03-29 at 14.02.59.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/0a7adf4e-d2cb-4e03-a012-d06ca2f94ad4/Screenshot%202023-03-29%20at%2014.02.59.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Support Forever code page.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code