Filter out duplicate tiers in comparison table (#2071)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Handle duplicate tiers in comparison table
- Add display names for comparison table (dog + cat)


![Screenshot 2023-03-30 at 16.54.09.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/d8b9e939-8bed-4c99-895d-0cf052b93343/Screenshot%202023-03-30%20at%2016.54.09.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Support pet insurance.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code