Pre-select cheapest offer (#2081)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Pre-select cheapest offer in the offer presenter

- Handle deductibles when remembering which option was selected after updating info

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Better UX.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code