Fix scroll in comparison table (#2099)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Fix an issue where you can't scroll comparison table (the done-button covers the content)


![Screenshot 2023-04-03 at 11.05.32.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/60d65227-a727-474e-b0c4-889599c38f76/Screenshot%202023-04-03%20at%2011.05.32.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code