Add size option to combobox (#2096)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Add size option to combobox according to design system

- Make nested mixed breed picker a small combobox


![Screenshot 2023-04-03 at 10.28.24.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/fe9788ac-f103-45c8-bfef-21fe67436404/Screenshot%202023-04-03%20at%2010.28.24.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code