Swap links in cart toast (#2103)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Swap links in the cart toast

Update copy in the cart toast


![Screenshot 2023-04-04 at 10.55.02.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/1654f6a4-e7ec-4112-80f3-ae0098d514ba/Screenshot%202023-04-04%20at%2010.55.02.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Design feedback

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code