Update how to get SSN to pass to Insurely (#2106)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fetch SSN from customer to pass to Insurely.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Before it was stored in the price intent data.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code