Fix content centering in fullscreen modal (#2107)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Fix centering of content in full screen modals

- Update boolean JSX props to be more explicit


![Screenshot 2023-04-04 at 15.08.03.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/dc993942-81f8-4e61-968c-23759221a818/Screenshot%202023-04-04%20at%2015.08.03.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Noticed that content was no longer centered in bank ID modals.

Boolean props in JSX will likely be deprecated in the future. This PR makes it more explicit what the props are doing.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code