Fix Storyblok visual editor setup (#2109)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Pass along query params from the API preview route

- Refactor cookie override to look the same as recommended by Storyblok

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

The visual editor is broken.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code