Read query params to initiate SB editor bridge (#2113)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Initiate SB editor brigde based on query params

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Noticed some cross-origin errors in the console with the previous solution.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code