Remove link to help page about deductibles (#2112)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.


![Screenshot 2023-04-05 at 15.07.06.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/598a1d1a-c0ae-49b0-9e5f-aaa17b2f93ca/Screenshot%202023-04-05%20at%2015.07.06.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Discussed in Pet stand-up.

## Jira issue(s): [GRW-2430]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-2430]: https://hedvig.atlassian.net/browse/GRW-2430?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ