Integrate C1 (customer support) script (#2141)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Integrate customer first (C1) script globally on the site.

Only show widget on-demand.

Remove intercom stuff.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We wanna test this out in staging to begin with.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code