Show correct price for selected offer in tier selector (#2149)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update to show the selected offer in tier selector instead of lowest price option.


![Screenshot 2023-04-13 at 11.19.56.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/c845c2fa-70b3-4c4f-b786-1a4de3eeb914/Screenshot%202023-04-13%20at%2011.19.56.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Reported as bug when the price doesn't match between tier selector and deductible selector.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code