Add reveal animation for campaign section (#2164)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add animation when expanding and collapsing the campaign section.

Checkout the preview app.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Testing it out with designers.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code