Make Insurely optional in Current Insurance picker (#2185)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Make it possible to edit current insurer without an Insurely client ID.


![Screenshot 2023-04-19 at 09.53.12.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/9434e86f-2f00-4235-b80e-fdc2faecbc04/Screenshot%202023-04-19%20at%2009.53.12.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We currently just skip that section if there is no Insurely client ID. However, it should still be possible to get help with cancellation even if we don't perform price matching.

Now we always let the user pick their insurer, and if there is an Insurely client ID we will perform price matching.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code