Add ability to replace a cart item (#2152)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Allow users to replace an entry in the cart.

Pick up the offer ID from the URL and use it to replace the entry in the cart.

Also use it to prefill some data in the form.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

To be able to add an "edit" button in the cart.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code