Use Peril.colorCode everywhere, warn about missing colors (should not happen in production, we would be able to check the logs now) (#2196)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

We've got colorCode values from PCMS now, use them instead of fallback

Added logs to track and missing values once we deploy this

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code