Show confirmation dialog before editing cart item (#2197)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Display a dialog with confirmation prompt after user clicks edit cart item

![localhost_8040_se_cart(iPhone SE).png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/052bbd1d-41d9-47ee-9d53-3778a64639c2/localhost_8040_se_cart(iPhone%20SE).png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Make it clear why you are sent back to the PDP

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code