Show preview in price calculator sections (#2213)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add support to show a preview label for each completed section in the price calculator.

Refactor the price calculator accordion slightly.


![Screenshot 2023-04-24 at 11.40.44.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/f4efb532-bde2-42c9-adfa-12c0bb55cb0b/Screenshot%202023-04-24%20at%2011.40.44.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Improve UX, especially when user needs to edit things about the price intent.

Just supports a single preview label for now. Let's discuss if we should support multiple labels in the future. Another discussion would be if we should move the form template to the backend.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code