Update English locale prefix (#2225)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Update English locale prefixes to `en-se`, `en-dk`, `en-no`

- Update locale utilities to match

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

- We wanna keep the same URL structure as the current site

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code