Update date formatting (#2226)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Use Lokalise translation when displaying "same day dates"

![Screenshot 2023-04-26 at 14.38.43.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/6990aad2-3b9c-4107-ace7-e61c13ff98d3/Screenshot%202023-04-26%20at%2014.38.43.png)

- Use consistent formatting of dates otherwise: `YYYY.MM.DD` (same for all locales)

![Screenshot 2023-04-26 at 14.48.50.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/09f86db5-bcb0-4027-89ca-081f7603f322/Screenshot%202023-04-26%20at%2014.48.50.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Design feedback.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code