Show tier and deductible in cart item details (#2223)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Display selected tier and deductible in cart item details

- Remove unused components

- Break out converter function between API and `CartEntry` object


![Screenshot 2023-04-25 at 17.21.51.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/b46038a8-b195-473a-b420-740875dea909/Screenshot%202023-04-25%20at%2017.21.51.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

I wanted to avoid passing down too much data to the visual components. This way we can keep the components as simple as possible and only pass down the data they need.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code