Add utility to auto format based on data field name (#2240)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Create utility function to autoformat values based on data field name

- Support ssn, registrationNumber, zipCode

![Screenshot 2023-04-28 at 11.32.00.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/67c5637d-6434-4346-b74b-b909ecb22fda/Screenshot%202023-04-28%20at%2011.32.00.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

- Consistent formatting of values without having to manually format them in every component

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code