Re-create price intent after adding to cart (#2279)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Create a new price intent after adding to cart and refresh price calculator


![Screenshot 2023-05-03 at 17.04.58.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/54bebc0f-dba9-4014-8abe-547295c89d70/Screenshot%202023-05-03%20at%2017.04.58.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We had a bug before where we didn't create a new price intent after adding to cart.

Instead we just removed the cookie and "refreshed" the page. But since it's now a static page nothing really happened. This is now fixed.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code