Change FALLBACK_LOCALE to sv-SE since this is what most visitors would want (#2311)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Change FALLBACK_LOCALE to sv-SE.  This is what we probably want by default for most visitors

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code