Add payment connect page (HOPE) (#2399)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add separate page to connect payment (existing members).

![Screenshot 2023-05-23 at 14.54.48.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/35e7179a-bbe5-4daa-be15-607d596e77fb/Screenshot%202023-05-23%20at%2014.54.48.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We will use this page to connect payment for existing members. The flow will begin in HOPE and will be redirected to this page.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code