Payment Connect: make it safe to reload page (#2417)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update payment connect page to require user to explicitly initiate the flow.


![localhost_8040_se_payment_connect_authorization_code=3ea88bec-e96b-4021-89c0-28eed73e0ca7(iPhone 12 Pro).png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/91d7f376-1054-4451-bc0f-79d049092b21/localhost_8040_se_payment_connect_authorization_code=3ea88bec-e96b-4021-89c0-28eed73e0ca7(iPhone%2012%20Pro).png)


![localhost_8040_se_payment_connect_authorization_code=3ea88bec-e96b-4021-89c0-28eed73e0ca7(iPhone 12 Pro) (1).png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/0110dc82-602e-43f7-b9df-987c46d4aaf4/localhost_8040_se_payment_connect_authorization_code=3ea88bec-e96b-4021-89c0-28eed73e0ca7(iPhone%2012%20Pro)%20(1).png)


![localhost_8040_se_payment_connect_authorization_code=9b9fa759-1bee-4de3-a289-e531d38b8fd6(iPhone 12 Pro).png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/040a6447-dd0f-4324-aec0-b5f504a6ff73/localhost_8040_se_payment_connect_authorization_code=9b9fa759-1bee-4de3-a289-e531d38b8fd6(iPhone%2012%20Pro).png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This makes it safe to send the payment link and for users to load the page any number of times. The authorization code is not used up until the user initiates the flow.

I'm not really sure what the design of the page should look like.

I'm not really sure if we should use a query parameter or a path parameter for the authorization code.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code