Cart toast: fix interact outside to dismiss (#2426)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix interact outside cart toast to dismiss it.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

There was an issue where the dialog content extended across the width of the page and the user could not interact with the page under it.


![Screenshot 2023-05-26 at 13.08.57.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/d751687f-337d-40b3-bb5e-c1b8069d5854/Screenshot%202023-05-26%20at%2013.08.57.png)

## Checklist before requesting a review

- [ ] I have performed a self-review of my code