[GEN-246] Car milage: display miles/year (#2427)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Display "miles/year" instead of "km/year" for car milage.


![Screenshot 2023-05-26 at 13.42.35.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/3ddefe5b-1fcf-4b5c-9e93-2e9977de9fa9/Screenshot%202023-05-26%20at%2013.42.35.png)


![Screenshot 2023-05-26 at 14.44.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/2c80ae8b-c917-41f6-84b7-b4bc13c7fd6b/Screenshot%202023-05-26%20at%2014.44.04.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This is the standard in Sweden so we translate both Swedish and English to "mil" instead of "km".

## Checklist before requesting a review

- [ ] I have performed a self-review of my code