[GEN-226] Support absolute URLs for the 'next' query param (#2425)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Support passing absolute URLs in the "next" query parameter.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This will be used by AdTraction for affiliate links.

We will still not allow any next redirects to go to a different domain than the current one.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code