Skip logging non-error (#2450)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This is not an error -- we only need to log it for debugging purposes. For example we get this if the user enters an invalid SSN.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code