[GEN-390] Add script to upload sourcemaps from Vercel (#2455)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add script to upload sourcemaps to Datadog from Vercel.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Vercel has a limit of amount of characters in the custom build script so we can't inline the full command.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code