[GEN-464] SelectInsuranceGrid: add option to open price calculator (#2502)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add CMS option to decide if links in "Select Insurance Grid" go to open price calculator or not.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

More flexibility for the CMS editor.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code