[GEN-400] Update to Insurely Blocks API (#2140)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Use the new "Blocks API" from Insurely.

- Get `insurelyConfigName` on PriceIntent.

- New environment variable for global Insurely customer ID


<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Insurely is discontinuing the old iFrame solution so only this new one will receive updates!

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code