Improve logging with Insurely (#2546)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add additional information to logs from Insurely integration

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Make it easier to see that it works as expected

## Checklist before requesting a review

- [ ] I have performed a self-review of my code