Group components in Storybook (#2555)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes
Storybook automatically groups components based on it's name and folder. So removing the title will group it under `components`. 

Let me know if this is better or not :D 

Before: 
<img width="766" alt="Screenshot 2023-06-15 at 11 35 29" src="https://github.com/HedvigInsurance/racoon/assets/6661511/0ae17dbf-2d22-449d-94f0-2941a462af44">

After:
<img width="652" alt="Screenshot 2023-06-15 at 11 35 03" src="https://github.com/HedvigInsurance/racoon/assets/6661511/96da0e65-6907-418e-92ac-014b336e70fd">

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Checklist before requesting a review

- [ ] I have performed a self-review of my code