[GEN-607] Pre-select price matched offers (#2628)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Pre-select offers that have been price matched in the offer presenter

- Update translations for when we are not able to match a competitor price

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We intend to control which offers are pre-selected (and/or highlighted) in the API, however, we are not able to do this yet. This is a temporary solution to launch Insurely for Car.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code