Extract Product Hero component from Purchase Form (#2630)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes


![Screenshot 2023-06-26 at 16.05.33.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/88fd943d-81fa-47bf-bacd-ae2f9cf63738/Screenshot%202023-06-26%20at%2016.05.33.png)


- Extract separate Product Hero component from Purchase Form

  - Based on Product Hero from Figma

- Integrate new component in Purchase Form

- Add Storybook stories for different variants

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

The Purchase Form is probably the most complex part of the code base and could use some refactoring. This is the first step in that direction. I plan to break it up more and more into smaller components.

## Checklist before requesting a review

- [ ] I have performed a self-review of my code